rename file a correctly
