+++ libs/common/src/admin-console/models/response/pre-validate-sponsorship.response.ts
